+++ src/utils/mixins.ts
@@ -32,18 +32,6 @@ export const Label = styled("label")<{ fullWidth?: boolean; left?: boolean }>(({
   marginRight: left ? theme.space.small : 0,
 }))
 
-export const labelTextHeight = 15
-
-export const LabelText = styled("span")(({ theme }) => ({
-  fontSize: theme.font.size.fineprint,
-  display: "block",
-  verticalAlign: "middle",
-  marginBottom: theme.space.base,
-  fontWeight: theme.font.weight.bold,
-  color: theme.color.text.lightest,
-  height: labelTextHeight,
-}))
-
 export const FormFieldControls = styled("div")({
   position: "absolute",
   top: 3,
